Pass runtime.firstmoduledata address from the client
The symbol is not available in non-dwarf + linux env.

+++ service/service_test.go
@@ -18,7 +18,7 @@ func TestAttachAndDetach(t *testing.T) {
 	tracer := &Tracer{}
 	args := AttachArgs{
 		Pid:                    cmd.Process.Pid,
-		InitialStartTracePoint: testutils.InfloopAddrMain,
+		InitialStartTracePoint: uintptr(testutils.InfloopAddrMain),
 		ProgramPath:            testutils.ProgramInfloop,
 		GoVersion:              runtime.Version(),
 	}
